3rd Fix Attempt
- Employed the downcasting to float32 from float64 the preprocessing pipeline
- Reduced max_generations to 15 from 30, also reduced rf parameters even further in hyperparam_bounds.
- Attempted to force fitness_score/generation run to n_jobs=1 (use single worker instead of parallel workers to reduce memory footprint)

- Added linear decay to the exploration components in loa & loa_vcs.py
- Added garbage collection calls in loa.py & loa_vcs.py

------
- Need to update comments to reflect changes.
- LOA.py now works without crashing but attempting to run all the metaheuristics still leads to a crash for the same reasons.

+++ Code/preprocessing.py
@@ -616,9 +616,9 @@ def load_processed_data():
 
     print(f"Loading data from {data_dir}...")
 
-    X_train = pd.read_csv(os.path.join(data_dir, "X_train.csv"))
-    X_val   = pd.read_csv(os.path.join(data_dir, "X_val.csv"))
-    X_test  = pd.read_csv(os.path.join(data_dir, "X_test.csv"))
+    X_train = pd.read_csv(os.path.join(data_dir, "X_train.csv"), dtype=np.float32)
+    X_val   = pd.read_csv(os.path.join(data_dir, "X_val.csv"),   dtype=np.float32)
+    X_test  = pd.read_csv(os.path.join(data_dir, "X_test.csv"),  dtype=np.float32)
     y_train = pd.read_csv(os.path.join(data_dir, "y_train.csv"))
     y_val   = pd.read_csv(os.path.join(data_dir, "y_val.csv"))
     y_test  = pd.read_csv(os.path.join(data_dir, "y_test.csv"))
